requirements.txt
wheel
cffi
fastapi==0.74.1
pydantic==1.9.0
PyJWT==2.3.0
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
uvicorn==0.17.5
passlib==1.7.4
python-dotenv==0.21.0
pymysql==1.0.2
requests
geopy
uuid
cryptography
urllib3==1.26.6 
pytz
bcrypt
pillow
python-multipart
kubernetes==29.0.0
beautifulsoup4
google-cloud-storage
schedule